feat(kgms-sdk,kgms,kgsdk): 筛选

+++ kgms-sdk/src/main/java/com/plantdata/kgcloud/sdk/rsp/app/explore/GisRelationRsp.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class GisRelationRsp extends GraphRelationRsp {
-    private Integer ruleId;
+    private Long ruleId;
 }
